Updated Navigation bar to conform to coding statndards for component directory configuration
